requirements.txt
altair
numpy
pandas
pydeck
streamlit
openai

langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
#streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
altair== 4.2.2
tiktoken
reportlab
pdf2image

